refactor(demo): Service Activities to headless namespace
