+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".BeginGetPasskeyActivity"
+                android:name=".headless.BeginGetPasskeyActivity"
                 android:excludeFromRecents="true"
                 android:exported="false">
             <intent-filter>
@@ -44,7 +44,7 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".GetPasskeyActivity"
+                android:name=".headless.GetPasskeyActivity"
                 android:excludeFromRecents="true"
                 android:exported="false">
             <intent-filter>
@@ -54,7 +54,7 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".CreatePasskeyActivity"
+                android:name=".headless.CreatePasskeyActivity"
                 android:excludeFromRecents="true"
                 android:exported="false">
             <intent-filter>
