+++ app/src/main/java/foundation/algorand/demo/headless/README.md
@@ -1,0 +1,3 @@
+# Overview
+
+Contains the Activities for the Credential Provider Service
