+++ app/src/main/res/layout/activity_begin_get_passkey.xml
@@ -5,6 +5,6 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".BeginGetPasskeyActivity">
+        tools:context=".headless.BeginGetPasskeyActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
