+++ app/src/main/res/layout/activity_get_passkey.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:tools="http://schemas.android.com/tools" xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:android="http://schemas.android.com/apk/res/android">
     <data>
-        <variable name="viewModel" type="foundation.algorand.demo.GetPasskeyViewModel"/>
+        <variable name="viewModel" type="foundation.algorand.demo.headless.GetPasskeyViewModel"/>
     </data>
 <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".GetPasskeyActivity"
+        tools:context=".headless.GetPasskeyActivity"
 >
 
     <TextView
